+++ universal-login-sdk/test/api/sdk.test.ts
@@ -1,15 +1,13 @@
 import chai, {expect} from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import sinonChai from 'sinon-chai';
-import {solidity, createFixtureLoader, deployContract} from 'ethereum-waffle';
-import {utils, providers, Wallet, Contract} from 'ethers';
+import {solidity, createFixtureLoader} from 'ethereum-waffle';
+import {providers, Wallet} from 'ethers';
 import {gnosisSafe} from '@unilogin/contracts';
-import {mockContracts} from '@unilogin/contracts/testutils';
-import {Message, GAS_BASE, PartialRequired, TEST_EXECUTION_OPTIONS, TEST_SDK_CONFIG, ETHER_NATIVE_TOKEN} from '@unilogin/commons';
+import {TEST_SDK_CONFIG} from '@unilogin/commons';
 import {RelayerUnderTest} from '@unilogin/relayer';
-import basicSDK, {transferMessage} from '../fixtures/basicSDK';
+import basicSDK from '../fixtures/basicSDK';
 import UniversalLoginSDK from '../../src/api/sdk';
-import {DeployedWallet} from '../../src';
 
 chai.use(solidity);
 chai.use(sinonChai);
@@ -24,17 +22,9 @@ describe('INT: SDK', () => {
   let sdk: UniversalLoginSDK;
   let contractAddress: string;
   let privateKey: string;
-  let wallet: Wallet;
-  let otherWallet: Wallet;
-  let message: PartialRequired<Message, 'from'>;
-  let walletContract: Contract;
-  let deployedWallet: DeployedWallet;
-  let ensName: string;
 
   beforeEach(async () => {
-    ({wallet, provider, otherWallet, sdk, privateKey, contractAddress, walletContract, relayer, ensName} = await loadFixture(basicSDK));
-    message = {...transferMessage, from: contractAddress, gasToken: ETHER_NATIVE_TOKEN.address, data: '0x'};
-    deployedWallet = new DeployedWallet(contractAddress, ensName, privateKey, sdk);
+    ({provider, sdk, privateKey, contractAddress, relayer} = await loadFixture(basicSDK));
   });
 
   afterEach(async () => {
@@ -58,57 +48,6 @@ describe('INT: SDK', () => {
         const response = sdk.getRelayerConfig();
         expect(response!.chainSpec.ensAddress).to.eq(expectedEnsAddress);
       });
-    });
-  });
-
-  describe('Execute signed message', () => {
-    it('Should execute signed message', async () => {
-      const expectedBalance = (await provider.getBalance(message.to!)).add(utils.parseEther('0.5'));
-      const {waitToBeSuccess} = await deployedWallet.execute(message);
-      const {transactionHash} = await waitToBeSuccess();
-      expect(transactionHash).to.match(/^[0x|0-9|a-f|A-F]{66}/);
-      expect(await provider.getBalance(message.to!)).to.eq(expectedBalance);
-    });
-
-    it('Should return transaction hash and proper state', async () => {
-      const {waitToBeSuccess} = await deployedWallet.execute(message);
-      const {transactionHash, state} = await waitToBeSuccess();
-      expect(transactionHash).to.be.properHex(64);
-      expect(state).to.eq('Success');
-    });
-
-    it('when not enough tokens ', async () => {
-      const mockToken = await deployContract(wallet, mockContracts.MockToken);
-      await mockToken.transfer(walletContract.address, 1);
-      message = {...message, gasToken: mockToken.address};
-      await expect(deployedWallet.execute(message)).to.be.eventually.rejectedWith('Not enough tokens');
-    });
-
-    it('when not enough ether', async () => {
-      const amountToTransfer = (await otherWallet.getBalance()).add(utils.parseEther('0.5'));
-      await expect(deployedWallet.execute({...message, to: otherWallet.address, value: amountToTransfer})).rejectedWith('Not enough tokens');
-    });
-
-    it('when not enough gas', async () => {
-      const baseGas = 88720;
-      const notEnoughGasLimit = 100;
-      message = {...message, gasLimit: baseGas + notEnoughGasLimit};
-      await expect(deployedWallet.execute(message)).to.be.eventually.rejectedWith(`Insufficient Gas. gasLimit should be greater than ${GAS_BASE}`);
-    });
-
-    it('Throws when the gas limit is above the relayers maxGasLimit', async () => {
-      const {sdk: secondSdk} = await loadFixture(basicSDK);
-      message.gasLimit = secondSdk.getRelayerConfig().maxGasLimit + 1;
-      const secondDeployedWallet = new DeployedWallet(contractAddress, '', privateKey, secondSdk);
-      await expect(secondDeployedWallet.execute(message)).to.be.eventually
-        .rejectedWith('Invalid gas limit. 500001 provided, when relayer\'s max gas limit is 500000');
-    });
-
-    it('Passes when the gas limit is equal to the relayers maxGasLimit', async () => {
-      const {sdk: secondSdk} = await loadFixture(basicSDK);
-      const secondDeployedWallet = new DeployedWallet(contractAddress, '', privateKey, secondSdk);
-      const {waitToBeSuccess} = await secondDeployedWallet.execute(message);
-      await expect(waitToBeSuccess()).to.be.eventually.fulfilled;
     });
   });
 
@@ -177,16 +116,6 @@ describe('INT: SDK', () => {
         expect(newDevicePrivateKey).to.be.properPrivateKey;
       });
     });
-  });
-
-  it('getMessageStatus', async () => {
-    await deployedWallet.addKey(otherWallet.address, TEST_EXECUTION_OPTIONS);
-    await deployedWallet.setRequiredSignatures(2, TEST_EXECUTION_OPTIONS);
-    const msg = {...message, to: otherWallet.address, nonce: await walletContract.nonce()};
-    const {messageStatus, waitForTransactionHash} = await deployedWallet.execute(msg);
-    const status = await sdk.getMessageStatus(messageStatus.messageHash);
-    expect(status.collectedSignatures.length).to.eq(1);
-    await waitForTransactionHash();
   });
 
   after(async () => {
