+++ universal-login-sdk/test/api/wallet/DeployedWallet.int.test.ts
@@ -2,12 +2,13 @@ import chai, {expect} from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import sinonChai from 'sinon-chai';
 import {createFixtureLoader, solidity} from 'ethereum-waffle';
-import {Contract} from 'ethers';
-import basicSDK from '../../fixtures/basicSDK';
+import {Contract, utils, providers, Wallet} from 'ethers';
+import basicSDK, {transferMessage} from '../../fixtures/basicSDK';
 import {RelayerUnderTest} from '@unilogin/relayer';
-import {walletFromBrain, DEFAULT_GAS_PRICE, createKeyPair, TEST_EXECUTION_OPTIONS} from '@unilogin/commons';
-import {DeployedWallet} from '../../../src';
+import {walletFromBrain, DEFAULT_GAS_PRICE, createKeyPair, TEST_EXECUTION_OPTIONS, Message, PartialRequired, deployContract, GAS_BASE, ETHER_NATIVE_TOKEN} from '@unilogin/commons';
+import UniversalLoginSDK, {DeployedWallet} from '../../../src';
 import {waitForSuccess} from '../../helpers/waitForSuccess';
+import {mockContracts} from '@unilogin/contracts/dist/esm/test';
 
 chai.use(solidity);
 chai.use(sinonChai);
@@ -18,17 +19,27 @@ const loadFixture = createFixtureLoader();
 const gasPrice = DEFAULT_GAS_PRICE;
 
 describe('INT: DeployedWallet', () => {
+  let provider: providers.Provider;
   let relayer: RelayerUnderTest;
   let mockToken: Contract;
   let deployedWallet: DeployedWallet;
   let ensName: string;
+  let message: PartialRequired<Message, 'from'>;
+  let wallet: Wallet;
+  let walletContract: Contract;
+  let otherWallet: Wallet;
+  let contractAddress: string;
+  let privateKey: string;
+  let sdk: UniversalLoginSDK;
   const publicKey = createKeyPair().publicKey;
   const publicKey2 = createKeyPair().publicKey;
 
   beforeEach(async () => {
+    // wallet, provider, otherWallet, walletContract,
     const {contractAddress, sdk, privateKey, ...rest} = await loadFixture(basicSDK) as any;
     ({relayer, mockToken, ensName} = rest);
     deployedWallet = new DeployedWallet(contractAddress, ensName, privateKey, sdk);
+    message = {...transferMessage, from: contractAddress, gasToken: ETHER_NATIVE_TOKEN.address, data: '0x'};
   });
 
   it('construction', async () => {
@@ -87,6 +98,67 @@ describe('INT: DeployedWallet', () => {
     expect(await deployedWallet.getNonce()).to.eq(1);
     expect(transactionHash).to.be.properHex(64);
     expect(state).to.eq('Success');
+  });
+
+  it('getMessageStatus', async () => {
+    await deployedWallet.addKey(otherWallet.address, TEST_EXECUTION_OPTIONS);
+    await deployedWallet.setRequiredSignatures(2, TEST_EXECUTION_OPTIONS);
+    const msg = {...message, to: otherWallet.address, nonce: await walletContract.nonce()};
+    const {messageStatus, waitForTransactionHash} = await deployedWallet.execute(msg);
+    const status = await sdk.getMessageStatus(messageStatus.messageHash);
+    expect(status.collectedSignatures.length).to.eq(1);
+    await waitForTransactionHash();
+  });
+
+  describe('Execute signed message', async () =>{
+    it('Should execute signed message', async () => {
+      const expectedBalance = (await provider.getBalance(message.to!)).add(utils.parseEther('0.5'));
+      const {waitToBeSuccess} = await deployedWallet.execute(message);
+      const {transactionHash} = await waitToBeSuccess();
+      expect(transactionHash).to.match(/^[0x|0-9|a-f|A-F]{66}/);
+      expect(await provider.getBalance(message.to!)).to.eq(expectedBalance);
+    });
+
+    it('Should return transaction hash and proper state', async () => {
+      const {waitToBeSuccess} = await deployedWallet.execute(message);
+      const {transactionHash, state} = await waitToBeSuccess();
+      expect(transactionHash).to.be.properHex(64);
+      expect(state).to.eq('Success');
+    });
+
+    it('when not enough tokens ', async () => {
+      const mockToken = await deployContract(wallet, mockContracts.MockToken);
+      await mockToken.transfer(walletContract.address, 1);
+      message = {...message, gasToken: mockToken.address};
+      await expect(deployedWallet.execute(message)).to.be.eventually.rejectedWith('Not enough tokens');
+    });
+
+    it('when not enough ether', async () => {
+      const amountToTransfer = (await otherWallet.getBalance()).add(utils.parseEther('0.5'));
+      await expect(deployedWallet.execute({...message, to: otherWallet.address, value: amountToTransfer})).rejectedWith('Not enough tokens');
+    });
+
+    it('when not enough gas', async () => {
+      const baseGas = 88720;
+      const notEnoughGasLimit = 100;
+      message = {...message, gasLimit: baseGas + notEnoughGasLimit};
+      await expect(deployedWallet.execute(message)).to.be.eventually.rejectedWith(`Insufficient Gas. gasLimit should be greater than ${GAS_BASE}`);
+    });
+
+    it('Throws when the gas limit is above the relayers maxGasLimit', async () => {
+      const {sdk: secondSdk} = await loadFixture(basicSDK);
+      message.gasLimit = secondSdk.getRelayerConfig().maxGasLimit + 1;
+      const secondDeployedWallet = new DeployedWallet(contractAddress, '', privateKey, secondSdk);
+      await expect(secondDeployedWallet.execute(message)).to.be.eventually
+        .rejectedWith('Invalid gas limit. 500001 provided, when relayer\'s max gas limit is 500000');
+    });
+
+    it('Passes when the gas limit is equal to the relayers maxGasLimit', async () => {
+      const {sdk: secondSdk} = await loadFixture(basicSDK);
+      const secondDeployedWallet = new DeployedWallet(contractAddress, '', privateKey, secondSdk);
+      const {waitToBeSuccess} = await secondDeployedWallet.execute(message);
+      await expect(waitToBeSuccess()).to.be.eventually.fulfilled;
+    });
   });
 
   afterEach(async () => {
